chore(MicrosoftClarity): move sdk loader to different file
