+++ src/integrations/MicrosoftClarity/browser.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import { NAME } from './constants';
 import Logger from '../../utils/logger';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
+import { loader } from './loader';
 
 const logger = new Logger(NAME);
 class MicrosoftClarity {
@@ -19,19 +19,7 @@ class MicrosoftClarity {
   }
 
   loadScript() {
-    (function (c, l, a, r, i, t, y) {
-      c[a] =
-        c[a] ||
-        function () {
-          (c[a].q = c[a].q || []).push(arguments);
-        };
-      t = l.createElement(r);
-      t.async = 1;
-      t.src = 'https://www.clarity.ms/tag/' + i;
-      t.setAttribute('data-loader', LOAD_ORIGIN);
-      y = l.getElementsByTagName(r)[0];
-      y.parentNode.insertBefore(t, y);
-    })(window, document, 'clarity', 'script', this.projectId);
+    loader(window, document, 'clarity', 'script', this.projectId);
     if (this.cookieConsent) {
       window.clarity('consent');
     }
